+++ lib/net.dart
@@ -1,7 +1,9 @@
 /// Internet related validators.
 ///
-/// E.g. email, ipv4, ipv6, mac, url.
+/// E.g. Email, IPv4, IPv6, MacAddress, Url.
 library net;
 
 export 'src/net/email.dart';
+export 'src/net/ipv4.dart';
 export 'src/net/req_email.dart';
+export 'src/net/req_ipv4.dart';